requirements.txt
asgiref==3.9.1
Django==4.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.9
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
